utility.pv: put integer value to channel without converting it to double

+++ applications/plugins/org.csstudio.utility.pv.epics/src/org/csstudio/utility/pv/epics/EPICS_V3_PV.java
@@ -480,7 +480,7 @@ public class EPICS_V3_PV
             }
             else if (new_value instanceof Integer)
             {
-                final double val = ((Integer)new_value).intValue();
+                final int val = ((Integer)new_value).intValue();
                 channel_ref.getChannel().put(val);
             }
             else if (new_value instanceof Integer [])
